# pkg/financial_dashboard_internal.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "financial_dashboard_internal.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "FinancialDashboard"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/payments/reload-overdraft-balance": {
      "post": {
        "summary": "History Avarage Balance",
        "operationId": "FinancialDashboard_ReloadOverdraftBalance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoROBRequest"
            }
          }
        ],
        "tags": [
          "FinancialDashboard"
        ]
      }
    },
    "/v1/payments/top-up/update-balance": {
      "post": {
        "summary": "Topup Balance",
        "operationId": "FinancialDashboard_TopupUpdateBalance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoTUBRequest"
            }
          }
        ],
        "tags": [
          "FinancialDashboard"
        ]
      }
    }
  },
  "definitions": {
    "protoROBRequest": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "number",
          "format": "double"
        },
        "clientIdList": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "protoResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      }
    },
    "protoTUBRequest": {
      "type": "object",
      "properties": {
        "client_id": {
          "type": "integer",
          "format": "int32"
        },
        "ref_id": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}